Rename a couple of ImgLib2 artifacts
As discussed at the 2014 October Madison ImgLib2 hackathon, the
imglib2-algorithms, imglib2-algorithms-fft, imglib2-algorithms-gpl
and imglib2-scripting artifact names were inconsistent with their
package prefixes. To avoid breaking backwards compatibility, the
artifacts were renamed to match the respective package prefixes.

The imglib2-ij artifact is also inconsistent with its package prefix,
but its prefix -- net.imglib2.img -- is also present in imglib2
core, and so the only "proper" solution would be to rename the
package, which would completely break backwards compatibility with
the relatively large amount of downstream code using imglib2-ij.
So we opted not to do that. (As an aside: Curtis's eventual tentative
plan for imglib2-ij is to merge it into imagej-legacy, which will be
a breaking change anyway. Best to avoid multiple breaking renames.)