POM: reorganize the authors list
The list is getting longer and longer. I was trying to keep it sorted by
git commit count, but that is a bad metric for a variety of reasons.
Furthermore, it is becoming increasingly annoying to compute, since
ImgLib2 is now split into a growing number of Git repositories

So let's use a simpler ordering: the three primary authors first,
followed by the rest of us riff-raff in alphabetical order. ;-)

This commit also adds Brian Northan to the list.

This change may take awhile to propagate all the way to the individual
project files's license headers and project licenses, but should happen
in January 2016 at the latest for the annual "Happy New Year" update.