Remove Dimiter Prodanov from copyright template
Actually, he did not write any code for ImgLib2.

The only relevant commit is:
https://github.com/imglib/imglib2/commit/a99c9e11a3a996368a940f1af8cde6174761add9

Which was done by Dimiter and me in collaboration. But that
code was only for benchmarking ImgLib2 and PixLib under
similar conditions. Really, the benchmark should have gone
into its own repository, with its own list of authors.

And since then, the PixLib portion of the benchmark was removed:
https://github.com/imglib/imglib2-tests/commit/b0910439d89fa35e8c2c29a1038fb8e0b45d35db